Add indent parameter in edn_format.dumps()
This adds a pretty printer similar to json.dumps(indent=<int>). However,
it does not follow Clojure formatting guidelines, instead formatting in
a way more common to users from other languages like Python.

So it will convert this:

    {"a": 1, "b": (1, 2, 3), "c": {"d": [1, 2, 3]}, "e": {1, 2, 3}}

To this:

    {
      :a 1
      :b (
        1
        2
        3
      )
      :c {
        :d [
          1
          2
          3
        ]
      }
      :e #{
        1
        2
        3
      }
    }

Instead of this:

    {
      :a 1
      :b (1 2 3)
      :c {
        :d [1 2 3]
      }
      :e #{1 2 3}
    }

This should be already better than the current status quo (that is, no
pretty printer at all).